Started conversion of graphics

+++ zeigen2/src/graphics.h
@@ -6,10 +6,10 @@
 
 void graphics_clear(Environment *environment);
 void graphics_present(Environment *environment);
-void graphics_render_at(Environment *environment, cairo_surface_t *surface, Int x, Int y);
-void graphics_render_centered_at(Environment *environment, cairo_surface_t *surface, Int x, Int y);
+void graphics_render_at(Environment *environment, cairo_surface_t *surface, Double x, Double y);
+void graphics_render_centered_at(Environment *environment, cairo_surface_t *surface, Double x, Double y);
 Bool graphics_render_at_position(Environment *environment, cairo_surface_t *surface, Value position);
-void graphics_fill(Environment *environment, Int red, Int green, Int blue, Int alpha);
+void graphics_fill(Environment *environment, Double red, Double green, Double blue, Double alpha);
 void graphics_calibrate(Environment *environment);
 
 #endif
